Move the abi to the aligned constants file
