+++ web/assets/js/react/constants/aligned.ts
@@ -139,6 +139,33 @@ export const leaderboardAbi = [
 	{ type: "error", name: "NoActiveParityGame", inputs: [] },
 ];
 
+export const zkArcadeNftAbi = [
+    {
+        type: "function",
+        stateMutability: "view",
+        name: "isWhitelisted",
+        inputs: [{ name: "user", type: "address" }],
+        outputs: [{ type: "bool" }],
+    },
+    {
+        type: "function",
+        stateMutability: "view",
+        name: "balanceOf",
+        inputs: [{ name: "owner", type: "address" }],
+        outputs: [{ type: "uint256" }],
+    },
+    {
+        type: "function",
+        stateMutability: "nonpayable",
+        name: "claimNFT",
+        inputs: [
+            { name: "merkleProof", type: "bytes32[]" },
+            { name: "tokenURI", type: "string" },
+        ],
+        outputs: [],
+    },
+] as const;
+
 export const eip712Domain = (
 	chainId: number,
 	batcherPaymentServiceAddress: Address
